DEV - Adding non-blocking synchronous methods recommendation service

+++ microservice/recommendation-service/src/main/java/felipe/microservices/core/recommendation/services/RecommendationServiceImpl.java
@@ -2,6 +2,8 @@ package felipe.microservices.core.recommendation.services;
 
 import java.util.List;
 
+import static java.util.logging.Level.FINE;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,6 +16,8 @@ import felipe.api.exceptions.InvalidInputException;
 import felipe.microservices.core.recommendation.persistence.RecommendationEntity;
 import felipe.microservices.core.recommendation.persistence.RecommendationRepository;
 import felipe.util.http.ServiceUtil;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
 @RestController
 public class RecommendationServiceImpl implements RecommendationService {
@@ -35,40 +39,52 @@ public class RecommendationServiceImpl implements RecommendationService {
   }
 
   @Override
-  public Recommendation createRecommendation(Recommendation body) {
-    try {
-      RecommendationEntity entity = mapper.apiToEntity(body);
-      RecommendationEntity newEntity = repository.save(entity);
-
-      LOG.debug("createRecommendation: created a recommendation entity: {}/{}", body.getProductId(),
-          body.getRecommendationId());
-      return mapper.entityToApi(newEntity);
-
-    } catch (DuplicateKeyException dke) {
-      throw new InvalidInputException(
-          "Duplicate key, Product Id: " + body.getProductId() + ", Recommendation Id:" + body.getRecommendationId());
+  public Mono<Recommendation> createRecommendation(Recommendation body) {
+    if(body.getProductId() < 1){
+      throw new InvalidInputException("Invalid productId: " + body.getProductId());
     }
+    
+    RecommendationEntity entity = mapper.apiToEntity(body);
+
+    Mono<Recommendation> newEntity = repository.save(entity)
+      .log(LOG.getName(), FINE)
+      .onErrorMap(DuplicateKeyException.class
+       ,ex -> new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Recommendation Id:" + body.getRecommendationId()))
+    .map(e -> mapper.entityToApi(e));
+
+    return newEntity;
+
+
   }
 
   @Override
-  public List<Recommendation> getRecommendations(int productId) {
+  public Flux<Recommendation> getRecommendations(int productId) {
 
     if (productId < 1) {
       throw new InvalidInputException("Invalid productId: " + productId);
     }
 
-    List<RecommendationEntity> entityList = repository.findByProductId(productId);
-    List<Recommendation> list = mapper.entityListToApiList(entityList);
-    list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
+    LOG.info("Will get recommendations for product with id={}", productId);
 
-    LOG.debug("getRecommendations: response size: {}", list.size());
+    return repository.findByProductId(productId)
+      .log(LOG.getName(),FINE)
+      .map(e -> mapper.entityToApi(e))
+      .map(e -> setServiceAddress(e));
 
-    return list;
   }
 
   @Override
-  public void deleteRecommendations(int productId) {
+  public Mono<Void> deleteRecommendations(int productId) {
+    if(productId > 1){
+      throw new InvalidInputException("Invalid productId: " + productId);
+    }
     LOG.debug("deleteRecommendations: tries to delete recommendations for the product with productId: {}", productId);
-    repository.deleteAll(repository.findByProductId(productId));
+
+    return repository.deleteAll(repository.findByProductId(productId));
+  }
+
+  private Recommendation setServiceAddress(Recommendation e) {
+    e.setServiceAddress(serviceUtil.getServiceAddress());
+    return e;
   }
 }
